Format expression service sources
