+++ Makefile
@@ -58,7 +58,7 @@ lint-fix:
 	@ $(UV_RUN) ruff check --preview --fix $(LINT_TARGETS)
 
 format:
-	@ $(UV_RUN) format $(LINT_TARGETS) 
+	@ $(UV_RUN) ruff format $(LINT_TARGETS) 
 
 typecheck:
 	@ $(UV_RUN) mypy $(LINT_TARGETS)
