+++ lizmap_server/expression_service/__init__.py
@@ -1,0 +1,1 @@
+from .service import ExpressionService as ExpressionService
