+++ lizmap_server/expression_service/models.py
@@ -1,3 +1,13 @@
+from typing import (
+    TypedDict,
+)
+
+
+class Body(TypedDict):
+    status: str
+    results: list
+    errors: list
+    features: int
 
 
 # Some expressions which can be evaluated on the server
@@ -10,4 +20,5 @@ ALLOWED_SAFE_EXPRESSIONS = {
     'represent_value',
     'round',
 }
+
 NOT_ALLOWED_EXPRESSION = "'not allowed'"
